Move luigi.hdfs to contrib and change imports
The change to test/minicluster.py is a terrible hack. But I'll have to
fix that in a later commit.

Changing the imports happen in this commit as well, otherwise the mock
magic in some of the tests do not work.
