+++ test/contrib/hdfs_test.py
@@ -25,7 +25,7 @@ import helpers
 import luigi
 import mock
 import luigi.format
-from luigi import hdfs
+from luigi.contrib import hdfs
 from luigi import six
 from minicluster import MiniClusterTestCase
 from nose.plugins.attrib import attr
@@ -38,7 +38,7 @@ class ComplexOldFormat(luigi.format.Format):
     """
 
     def hdfs_writer(self, output_pipe):
-        return self.pipe_writer(luigi.hdfs.Plain.hdfs_writer(output_pipe))
+        return self.pipe_writer(luigi.contrib.hdfs.Plain.hdfs_writer(output_pipe))
 
     def pipe_writer(self, output_pipe):
         return luigi.format.UTF8.pipe_writer(output_pipe)
@@ -696,9 +696,9 @@ class HdfsClientTest(MiniClusterTestCase):
         self.assertEqual(4, len(entries[5]), msg="%r" % entries)
         self.assertEqual(path + '/sub2/file4.dat', entries[5][0], msg="%r" % entries)
 
-    @mock.patch('luigi.hdfs.call_check')
+    @mock.patch('luigi.contrib.hdfs.call_check')
     def test_cdh3_client(self, call_check):
-        cdh3_client = luigi.hdfs.HdfsClientCdh3()
+        cdh3_client = luigi.contrib.hdfs.HdfsClientCdh3()
         cdh3_client.remove("/some/path/here")
         self.assertEqual(['fs', '-rmr', '/some/path/here'], call_check.call_args[0][0][-3:])
 
@@ -715,7 +715,7 @@ class HdfsClientTest(MiniClusterTestCase):
         preturn.communicate = comm
         popen.return_value = preturn
 
-        apache_client = luigi.hdfs.HdfsClientApache1()
+        apache_client = luigi.contrib.hdfs.HdfsClientApache1()
         returned = apache_client.exists("/some/path/somewhere")
         self.assertTrue(returned)
 
@@ -724,7 +724,7 @@ class HdfsClientTest(MiniClusterTestCase):
         self.assertFalse(returned)
 
         preturn.returncode = 13
-        self.assertRaises(luigi.hdfs.HDFSCliError, apache_client.exists, "/some/path/somewhere")
+        self.assertRaises(luigi.contrib.hdfs.HDFSCliError, apache_client.exists, "/some/path/somewhere")
 
 
 class SnakebiteConfigTest(unittest.TestCase):
@@ -747,7 +747,7 @@ class _MiscOperationsMixin(object):
 
     def get_target(self):
         fn = '/tmp/foo-%09d' % random.randint(0, 999999999)
-        t = luigi.hdfs.HdfsTarget(fn)
+        t = luigi.contrib.hdfs.HdfsTarget(fn)
         with t.open('w') as f:
             f.write('test')
         return t
@@ -770,7 +770,7 @@ class _MiscOperationsMixin(object):
 @attr('minicluster')
 class TestCliMisc(MiniClusterTestCase, _MiscOperationsMixin):
     def get_client(self):
-        return luigi.hdfs.create_hadoopcli_client()
+        return luigi.contrib.hdfs.create_hadoopcli_client()
 
 
 @attr('minicluster')
@@ -779,4 +779,4 @@ class TestSnakebiteMisc(MiniClusterTestCase, _MiscOperationsMixin):
         if six.PY3:
             raise unittest.SkipTest("snakebite doesn't work on Python yet.")
 
-        return luigi.hdfs.SnakebiteHdfsClient()
+        return luigi.contrib.hdfs.SnakebiteHdfsClient()
